Ditched microtime.

+++ dev/db/schema.sql
@@ -6,7 +6,7 @@ BEGIN TRANSACTION;
 -- Use for logging access and applying rate limiting.
 CREATE TABLE IF NOT EXISTS sys_accesslog (
     id            INTEGER  PRIMARY KEY AUTOINCREMENT,
-    accessTime    REAL     NOT NULL,
+    accessTime    INTEGER  NOT NULL,
     clientHash    TEXT     NOT NULL,
     apiRequest    TEXT     NOT NULL
 );
